% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Test function}
\usage{
hello(m = "(smth else)")
}
\arguments{
\item{m}{it's an addition to the message}
}
\value{
just prints, nothing to return
}
\description{
It has some arguments. Let's present them.
}
\examples{
hello()
}
